Storing/Declaring big numbers

+++ Variables/Example3.java
@@ -1,6 +1,6 @@
 public class StoringBigInt {
     public static void main(String[] args) {
-        int Int_Big_Number = (int) 1_000_000_000;
+        int Int_Big_Number = 1_000_000_000;
         System.out.println("Printing Big Number stored in an INT: " + Int_Big_Number);
         
         long Long_Big_Number = 1_000_000_000;
